+++ data/blood/ketone/ketone.go
@@ -12,6 +12,8 @@ const (
 
 	MmolLMinimum = 0.0
 	MmolLMaximum = 10.0
+
+	MmolLPrecisionFactor float64 = 100000.0
 )
 
 func Units() []string {
@@ -39,5 +41,22 @@ func NormalizeUnits(units *string) *string {
 }
 
 func NormalizeValueForUnits(value *float64, units *string) *float64 {
+	if value != nil && units != nil {
+		switch *units {
+		case MmolL, Mmoll:
+			// TODO: Normalize mmol/L values to standard precision
+			// return NormalizePrecisionForUnits(value, units)
+		}
+	}
+	return value
+}
+
+func NormalizePrecisionForUnits(value *float64, units *string) *float64 {
+	if value != nil && units != nil {
+		switch *units {
+		case MmolL, Mmoll:
+			return pointer.Float64(float64(int(*value*MmolLPrecisionFactor+0.5)) / MmolLPrecisionFactor)
+		}
+	}
 	return value
 }